# Port Intel optimizations (adler32, chunkcopy) to cloudflare