--- a/adler32.c
+++ b/adler32.c
@@ -61,7 +61,7 @@ local uLong adler32_combine_ OF((uLong adler1, uLong adler2, z_off64_t len2));
 #  define MOD63(a) a %= BASE
 #endif
 
-#if defined(ADLER32_SIMD_NEON)
+#if defined(ADLER32_SIMD_NEON) || defined (ADLER32_SIMD_SSSE3)
 #include "adler32_simd.h"
 #endif
 
@@ -74,7 +74,7 @@ uLong ZEXPORT adler32(adler, buf, len)
     unsigned long sum2;
     unsigned n;
 
-#if defined(ADLER32_SIMD_NEON)
+#if defined(ADLER32_SIMD_NEON) || defined(ADLER32_SIMD_SSSE3)
     if (buf && len >= 64)
         return adler32_simd_(adler, buf, len);
 #endif