This series ports the x86 inflate performance improvements from the chromium fork of zlib.

Based on 3 patches:
* Noel Gordon (noel@chromium.org)
   17bbb3d73c84 - "zlib adler_simd.c"
* Adenilson Cavalcanti (adenilson.cavalcanti@arm.com)
   3060dcb - "zlib: inflate using wider loads and stores"
* Noel Gordon (noel@chromium.org)
   64ffef0 - "Improve zlib inflate speed by using SSE2 chunk copy

With these changes, the inflate performance improvement is 15-35% when tested with a modified zpipe.c and the Silesia corpus.