@janaknat 2020-09-28
Performance numbers using the Silesia Corpus and a modified zpipe.c

Dataset| File | Throughput before (kb/ms)| Throughput after (kb/ms)| % change |
--------|-----|--------------------|----------------|---|
Silesia | xml | 65.21 | 84.29| +29%
 | | x-ray | 138.18 | 159.91| +15%
 | | webster | 75.03 | 95.93| +26%
 | | sao | 155.98 | 189.81 | +21%
 | | samba | 88.70 | 111.74 | +26%
 | | reymont | 73.50 | 98.45 | +34%
 | | osdb | 99.36 | 120.32 | +21%
 | | ooffice | 98.12 | 118.96 | +20%
 | | nci | 55.70 | 69.41 | +25%
 | | mr | 92.76 | 115.98 | +25%
 | | mozilla | 101.37 | 121.30 | +19%
 | | dickens | 81.98 | 109.82 | +34%