@RJVB 2020-09-29
So I built and compared #82035d0687c8d34981c93589f35982149fb34591 with #836eb111a5c5df7db3f2469a867a6f7c1b2e7bdb by running minigzip on a 305Mb big source tarball (QtBase 5.12.6, uncompressed; no options, reading from stdin and sending the output to /dev/null).

To my surprise, compression time increased about 5x in the current release. Please tell me that makes sense because the compression settings were changed?!