@janaknat 2020-09-29
@RJVB Can you provide a link to the source tarball? How did you measure the time taken?

Also, the changes made in this PR are for de-compression(inflate). No changes were made to the compression code path.