@vkrasnov 2020-09-29
This is very weird, but he is right, in fact inflate is slower too, because somehow the -O3 flags goes missing after running configure.