@vkrasnov 2020-09-29
Fixed now in https://github.com/cloudflare/zlib/pull/24