@RJVB 2020-09-29
>Thanks @RJVB !

Well, thank *you* for finding the cause so quickly!

I'm still at almost 2x slower compression (and that is in fact compared against a 32-bit build of the older code!) but decompression is indeed significantly faster. I don't know if you can compare the number of instructions between 64 bit and 32 bit builds but according to `perf stat` the older version requires more than 2x as many instructions to decompress the same file (and almost 2x as many cycles).