@janaknat 2020-10-02
@RJVB Do you see the same compression slow-down without the changes in this PR?