@neurolabusc 2020-10-02
@janaknat I do not see any benefits for [compressing Silesia files to/from gzip format](https://github.com/neurolabusc/zlib-bench-python). I compared the latest version of this repository (zlibCFX) to a [fork from June (zlibCF)](https://github.com/rordenlab/zlib) on my MacBook with in-built SSD. Same performance regardless of whether I use make or cmake to compile the application. I suspect this is because Gzip uses crc32 not Adler-32.

On the other hand, decompression is faster, but only appreciably if compiled with `./configure  --static; make` which reports the `DINFLATE_CHUNK_SIMD_SSE2` option. @janaknat, since you are familiar with these new features, I would be grateful if you would consider updating cmake to support them. 

|CompressMethod|Level|Min|Mean|Max|mb/s|%|
|-|-|-|-|-|-|-|
|zlibCF|1|2350|2405|2483|90|35.80|
|zlibCF|2|2472|2482|2498|86|35.00|
|zlibCF|3|2682|2736|2802|79|34.37|
|zlibCF|4|2907|3072|3248|73|33.46|
|zlibCF|5|3585|4071|4497|59|32.61|
|zlibCF|6|4460|4860|5542|48|32.35|
|zlibCF|7|5008|5459|5806|42|32.25|
|zlibCF|8|7400|7551|7856|29|32.17|
|zlibCF|9|10099|10678|11507|21|32.15|
|zlibCFX|1|2524|2584|2678|84|35.80|
|zlibCFX|2|2542|2624|2761|83|35.00|
|zlibCFX|3|2772|2899|3000|76|34.37|
|zlibCFX|4|2891|3141|3443|73|33.46|
|zlibCFX|5|3688|3758|3928|57|32.61|
|zlibCFX|6|4475|4678|5103|47|32.35|
|zlibCFX|7|5193|5403|5980|41|32.25|
|zlibCFX|8|7498|7671|7996|28|32.17|
|zlibCFX|9|10093|10274|10558|21|32.15|

|DecompressMethod|Min|Mean|Max|mb/s|
|-|-|-|-|-|
|zlibCF|42683|43312|43883|134.07|
|zlibCFX|41457|42038|42833|138.03|
|zlibCF(make)|36631|36814|37127|156.22|