@RJVB 2020-10-02
>@RJVB Do you see the same compression slow-down without the changes in this PR?

I'd have to check - I didn't yet because I never thought it was due to this change esp. after learning that the optimisations are only for decompression.