@janaknat 2020-10-05
@neurolabusc These are de-compression optimization changes, so not surprising there are any changes in compression performance. 

I can take a look at the work that needs to be done for cmake.